Move out logger/sink in newrelic
